// backend/question_service/src/firebase/interface.ts
import { UUID } from 'crypto';

export interface Question {
  question_id?: UUID; // UUID is generated by firebase and returned automatically
  title: string;
  category: string;
  complexity: number;
  link: string;
  description: string;
}

export interface QuestionData {
  title?: string;
  category?: string;
  complexity?: number;
  link?: string;
  description?: string;
  [key: string]: any; // This allows any key-value pair
}